# docs/openapi/all-services.yaml
openapi: 3.1.0
info:
  title: BeckLab Tools — Full API Reference
  version: 1.0.0
  description: Complete API reference for all BeckLab micro-services. Each service is independently accessible at its own subdomain.
  contact:
    name: BeckLab
    url: https://tools.becklab.cloud
servers:
  - url: https://hash.tools.becklab.cloud
    description: Hash Service
  - url: https://short.tools.becklab.cloud
    description: URL Shortener
  - url: https://convert.tools.becklab.cloud
    description: Base64/Converter
  - url: https://cron.tools.becklab.cloud
    description: Cron Jobs
  - url: https://dns.tools.becklab.cloud
    description: DNS Monitor
  - url: https://webhook.tools.becklab.cloud
    description: Webhook Relay
  - url: https://fmt.tools.becklab.cloud
    description: YAML/JSON Tool
  - url: https://qr.tools.becklab.cloud
    description: QR Generator
  - url: https://editor.tools.becklab.cloud
    description: Image Editor
  - url: https://dashboard.tools.becklab.cloud
    description: BeckFlow Gateway
  - url: https://auth.tools.becklab.cloud
    description: Auth Service
...
paths:
...
  /api/v1/dns:
    post:
      summary: Perform DNS lookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain]
              properties:
                domain: { type: string, example: "example.com" }
                recordType:
                  type: string
                  enum: [A, AAAA, MX, TXT, CNAME, NS, SOA]
                  default: A
      responses:
        '200': { description: DNS lookup results }
...
  /api/v1/tls:
    post:
      summary: Check TLS certificate details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain]
              properties:
                domain: { type: string, example: "example.com" }
                port: { type: integer, default: 443 }
      responses:
        '200': { description: TLS certificate details }
  # === Webhook Relay ===
...
  /api/v1/editor/save:
    post:
      summary: Save processed image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data: { type: string, description: Base64-encoded image }
                filename: { type: string }
      responses:
        '200': { description: Image saved }
  # === BeckFlow Gateway ===
...